ip_validator/requirements.txt
geoip2
fastapi
pytest
uvicorn
pytest-asyncio
grpcio
grpcio-tools